issue #399
