+++ src/util/src/tsocket.c
@@ -23,17 +23,14 @@
 #include <pthread.h>
 #include <stdarg.h>
 #include <sys/types.h>
-#include <sys/utsname.h>
 
 #include "os.h"
 #include "tglobalcfg.h"
 #include "tlog.h"
 #include "tsocket.h"
 #include "tutil.h"
-#include "tsystem.h"
 
 unsigned int ip2uint(const char *const ip_addr);
-bool isRunningWSLv1();
 
 /*
  * Function to get the public ip address of current machine. If get IP
@@ -306,7 +303,7 @@ int taosOpenUdpSocket(char *ip, short port) {
   if (taosSetSockOpt(sockFd, SOL_SOCKET, SO_NO_CHECK, (void *)&nocheck, sizeof(nocheck)) < 0) {
     // no_check is not implemented in WSL
     // skip the following check if system running WSLv1
-    if (!isRunningWSLv1()) {
+    if (!taosIsRunningWSLv1()) {
       pError("setsockopt SO_NO_CHECK failed: %d (%s)", errno, strerror(errno));
       close(sockFd);
       return -1;
@@ -555,20 +552,4 @@ int taosCopyFds(int sfd, int dfd, int64_t len) {
   }
 
   return 0;
-}
-
-// check if the linux running is WSL
-bool isRunningWSLv1() {
-  struct utsname buf;
-  if (uname(&buf)) {
-    pPrint(" can't fetch os info");
-    return false;
-  }
-
-  if (strstr(buf.release, "Microsoft") != 0) {
-    pPrint(" using WSLv1");
-    return true;
-  }
-
-  return false;
 }
